+++ src/index.rs
@@ -6,62 +6,35 @@
 //! * [`RowMajor`]: Elements in the same row are contiguous in memory.
 //! * [`ColMajor`]: Elements in the same column are contiguous in memory.
 //!
-//! [`Index`] is a wrapper around a `usize` that is tagged with a specific layout and position type.
-//!
 //! # Examples
 //!
 //! ```rust
-//! use ixy::{Pos, index::{Index, RowMajor, ColMajor, Layout}};
+//! use ixy::{Pos, index::{RowMajor, ColMajor, Layout}};
 //!
 //! let pos = Pos::new(2, 3);
 //! let width = 5;
 //!
 //! // Convert a 2D position to a 1D index in row-major layout
-//! let index_row = RowMajor::to_1d(pos, width);
-//! assert_eq!(index_row.index, 17); // 3 * 5 + 2
+//! let index = RowMajor::to_1d(pos, width);
+//! assert_eq!(index, 17); // 3 * 5 + 2
 //!
 //! // Convert a 1D index back to a 2D position in row-major layout
-//! let pos_row = RowMajor::to_2d(index_row, width);
-//! assert_eq!(pos_row, Pos::new(2, 3));
+//! let pos: Pos<i32> = RowMajor::to_2d(index, width);
+//! assert_eq!(pos, Pos::new(2, 3));
 //! ```
 
 use core::marker::PhantomData;
 
 use crate::{Pos, Rect, int::Int};
 
-/// A linear memory index that can be used to access elements in a 2-dimensional array.
-#[repr(transparent)]
-pub struct Index<T: Int, L: Layout = RowMajor> {
-    pub index: usize,
-    _pos_type: PhantomData<T>,
-    _layout: PhantomData<L>,
-}
-
-impl<T: Int, L: Layout> Index<T, L> {
-    /// Wraps a `usize` index into an `Index`.
-    ///
-    /// The generics `T` and `L` are used to specify the integer type and layout of the index.
-    #[must_use]
-    pub const fn new(index: usize) -> Self {
-        Self {
-            index,
-            _pos_type: PhantomData,
-            _layout: PhantomData,
-        }
-    }
-}
-
 /// A mapping between a 2-dimensional array and a linear memory layout.
 #[allow(private_bounds)]
 pub trait Layout: Sized + crate::internal::Sealed {
-    /// Whether the layout is row-major.
-    const IS_ROW_MAJOR: bool;
-
     /// Converts a 2-dimensional position to a linear memory index.
-    fn to_1d<T: Int>(pos: Pos<T>, width: usize) -> Index<T, Self>;
+    fn to_1d<T: Int>(pos: Pos<T>, width: usize) -> usize;
 
     /// Converts a linear memory index to a 2-dimensional position.
-    fn to_2d<T: Int>(index: Index<T, Self>, width: usize) -> Pos<T>;
+    fn to_2d<T: Int>(index: usize, width: usize) -> Pos<T>;
 
     /// Creates an iterator of the given bounds, yielding positions in the order defined.
     fn iter_pos<T: Int>(bounds: Rect<T>) -> impl Iterator<Item = Pos<T>>;
@@ -121,20 +94,19 @@ where
 }
 
 /// Each row is stored contiguously in memory, with the first row at the lowest address.
-pub struct RowMajor;
+pub enum RowMajor {}
 
 impl crate::internal::Sealed for RowMajor {}
-impl Layout for RowMajor {
-    const IS_ROW_MAJOR: bool = true;
 
-    fn to_1d<T: Int>(pos: Pos<T>, width: usize) -> Index<T, Self> {
-        Index::new(pos.y.to_usize() * width + pos.x.to_usize())
+impl Layout for RowMajor {
+    fn to_1d<T: Int>(pos: Pos<T>, width: usize) -> usize {
+        pos.y.to_usize() * width + pos.x.to_usize()
     }
 
-    fn to_2d<T: Int>(index: Index<T, Self>, width: usize) -> Pos<T> {
+    fn to_2d<T: Int>(index: usize, width: usize) -> Pos<T> {
         Pos {
-            x: T::from_usize(index.index % width),
-            y: T::from_usize(index.index / width),
+            x: T::from_usize(index % width),
+            y: T::from_usize(index / width),
         }
     }
 
@@ -148,24 +120,19 @@ impl Layout for RowMajor {
 }
 
 /// Each column is stored contiguously in memory, with the first column at the lowest address.
-pub struct ColMajor;
+pub enum ColMajor {}
 
 impl crate::internal::Sealed for ColMajor {}
-impl Layout for ColMajor {
-    const IS_ROW_MAJOR: bool = false;
 
-    fn to_1d<T: Int>(pos: Pos<T>, width: usize) -> Index<T, Self> {
-        Index {
-            index: (pos.x.to_usize() * width + pos.y.to_usize()),
-            _pos_type: PhantomData,
-            _layout: PhantomData,
-        }
+impl Layout for ColMajor {
+    fn to_1d<T: Int>(pos: Pos<T>, width: usize) -> usize {
+        pos.x.to_usize() * width + pos.y.to_usize()
     }
 
-    fn to_2d<T: Int>(index: Index<T, Self>, width: usize) -> Pos<T> {
+    fn to_2d<T: Int>(index: usize, width: usize) -> Pos<T> {
         Pos {
-            x: T::from_usize(index.index / width),
-            y: T::from_usize(index.index % width),
+            x: T::from_usize(index / width),
+            y: T::from_usize(index % width),
         }
     }
 
@@ -182,9 +149,8 @@ impl Layout for ColMajor {
 mod tests {
     extern crate alloc;
 
-    use crate::pos;
-
     use super::*;
+    use crate::pos;
     use alloc::vec::Vec;
 
     #[test]
@@ -193,7 +159,7 @@ mod tests {
         let width = 5;
 
         let index = RowMajor::to_1d(pos, width);
-        assert_eq!(index.index, 17); // 3 * 5 + 2
+        assert_eq!(index, 17); // 3 * 5 + 2
     }
 
     #[test]
@@ -202,25 +168,25 @@ mod tests {
         let width = 5;
 
         let index = ColMajor::to_1d(pos, width);
-        assert_eq!(index.index, 13); // 2 * 5 + 3
+        assert_eq!(index, 13); // 2 * 5 + 3
     }
 
     #[test]
     fn from_row_major() {
-        let index = Index::<i32, RowMajor>::new(17);
+        let index = 17;
         let width = 5;
 
-        let pos = RowMajor::to_2d(index, width);
+        let pos: Pos<i32> = RowMajor::to_2d(index, width);
         assert_eq!(pos.x, 2); // 17 % 5
         assert_eq!(pos.y, 3); // 17 / 5
     }
 
     #[test]
     fn from_col_major() {
-        let index = Index::<i32, ColMajor>::new(13);
+        let index = 13;
         let width = 5;
 
-        let pos = ColMajor::to_2d(index, width);
+        let pos: Pos<i32> = ColMajor::to_2d(index, width);
         assert_eq!(pos.x, 2); // 13 / 5
         assert_eq!(pos.y, 3); // 13 % 5
     }