+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2025-07-19
+
+### Removed
+
+- `Index`
+- `Layout::{to_1d|to_2d}` now return (and use) `usize` instead of `Index`
+- `Layout::IS_ROW_MAJOR`
+
 ## [0.4.0] - 2025-07-18
 
-## Removed
+### Removed
 
 - `Rect::from_ltwh_unsigned`
 - `int::UnsignedInt`, which is now unused